refactor: refactor code for restClientPage
