+++ app/routes/privateRoutes/restClientPage/ServerFetchPanel.tsx
@@ -14,9 +14,9 @@ import styles from './RestClientPage.module.css';
 import { Response } from './components/response';
 import { parserList } from './components/response/parserList';
 
-export { serverFetchLoader as loader } from './serverFetchLoader';
+export { serverFetchLoader as loader } from './loaders/serverFetchLoader';
 
-export default function ServerFetch({ loaderData }: Route.ComponentProps) {
+export default function ServerFetchPanel({ loaderData }: Route.ComponentProps) {
   const { data, error, status } = (loaderData || {}) as LoaderData;
   const { setParser } = useActions();
 
@@ -27,6 +27,7 @@ export default function ServerFetch({ loaderData }: Route.ComponentProps) {
   return (
     <div className={styles.response}>
       <h3>{serverFetch.title}</h3>
+
       <div className={styles.parser}>
         <p className={styles.title}>{serverMessages.selectParser}</p>
