+++ app/routes/privateRoutes/restClientPage/components/headers/Headers.module.css
@@ -13,7 +13,7 @@
   th,
   td {
     border: var(--spacing-1) solid var(--md-sys-color-surface-bright);
-    padding: var(--spacing-5) var(--spacing-4);
+    padding: var(--spacing-4);
     text-align: center;
     vertical-align: middle;
   }
