+++ app/routes/privateRoutes/restClientPage/components/urlBox/index.tsx
@@ -7,13 +7,7 @@ import {
 import { useSelector } from 'react-redux';
 import { useSearchParams } from 'react-router';
 
-import {
-  HttpMethods,
-  InputID,
-  InputType,
-  Protocols,
-  SearchParams,
-} from '@/sources/enums';
+import { HttpMethods, InputID, InputType, Protocols } from '@/sources/enums';
 
 import { buttons as buttonMessages } from '@/sources/messages/buttons';
 
@@ -24,7 +18,12 @@ import { Select } from '@/components/ui/select';
 import { useActions } from '@/utils/hooks/useActions';
 
 import styles from './UrlBox.module.css';
-import { handleMethod, handleProtocol, handleUrl } from './handlers';
+import {
+  handleMethod,
+  handleProtocol,
+  handleServerFetch,
+  handleUrl,
+} from './handlers';
 import { methodList } from './lists/methodList';
 import { protocolList } from './lists/protocolList';
 
@@ -34,23 +33,7 @@ export const UrlBox = () => {
   const protocol = useSelector(selectProtocol);
 
   const { setMethod, setProtocol, setUrl } = useActions();
-
   const setSearchParams = useSearchParams()[1];
-
-  const handleServerFetch = async () => {
-    if (!url) return;
-
-    try {
-      const newSearchParams = new URLSearchParams();
-      newSearchParams.set(SearchParams.URL, url);
-      newSearchParams.set(SearchParams.METHOD, method);
-      newSearchParams.set(SearchParams.PROTOCOL, protocol);
-
-      setSearchParams(newSearchParams);
-    } catch (error) {
-      console.log('Error:', error);
-    }
-  };
 
   return (
     <div className={styles.container}>
@@ -74,7 +57,13 @@ export const UrlBox = () => {
         setInput={value => handleUrl(value, setUrl)}
       />
 
-      <Button onClick={handleServerFetch}>{buttonMessages.send}</Button>
+      <Button
+        onClick={() =>
+          handleServerFetch(url, method, protocol, setSearchParams)
+        }
+      >
+        {buttonMessages.send}
+      </Button>
     </div>
   );
 };
